+++ siso-ios/Matching/Views/MatchingCardView.swift
@@ -44,7 +44,7 @@ public struct MatchingCardView: View {
             Spacer()
         }
         .background {
-            backgroundView
+           // backgroundView
         }
         
     }
